Reformat + extract glyph tests
