+++ sparse_strips/vello_cpu/tests/basic.rs
@@ -3,7 +3,10 @@
 
 //! Tests for basic functionality.
 
-use crate::util::{bevel_stroke_2, check_ref, circular_star, crossed_line_star, get_ctx, miter_stroke_2, render_pixmap};
+use crate::util::{
+    bevel_stroke_2, check_ref, circular_star, crossed_line_star, get_ctx, miter_stroke_2,
+    render_pixmap,
+};
 use skrifa::MetadataProvider;
 use skrifa::raw::FileRef;
 use std::f64::consts::PI;
@@ -15,7 +18,6 @@ use vello_common::glyph::Glyph;
 use vello_common::kurbo::{Affine, BezPath, Circle, Join, Point, Rect, Shape, Stroke, Vec2};
 use vello_common::peniko::{self, Blob, Compose, Fill, Font};
 use vello_cpu::RenderContext;
-
 
 #[test]
 fn empty_1x1() {
@@ -513,121 +515,6 @@ fn filled_vertical_hairline_rect_2() {
 }
 
 #[test]
-fn filled_glyphs() {
-    let mut ctx = get_ctx(300, 70, false);
-    let font_size: f32 = 50_f32;
-    let (font, glyphs) = layout_glyphs("Hello, world!", font_size);
-
-    ctx.set_transform(Affine::translate((0., f64::from(font_size))));
-    ctx.set_paint(REBECCA_PURPLE.with_alpha(0.5).into());
-    ctx.glyph_run(&font)
-        .font_size(font_size)
-        .fill_glyphs(glyphs.into_iter());
-
-    check_ref(&ctx, "filled_glyphs");
-}
-
-#[test]
-fn stroked_glyphs() {
-    let mut ctx = get_ctx(300, 70, false);
-    let font_size: f32 = 50_f32;
-    let (font, glyphs) = layout_glyphs("Hello, world!", font_size);
-
-    ctx.set_transform(Affine::translate((0., f64::from(font_size))));
-    ctx.set_paint(REBECCA_PURPLE.with_alpha(0.5).into());
-    ctx.glyph_run(&font)
-        .font_size(font_size)
-        .stroke_glyphs(glyphs.into_iter());
-
-    check_ref(&ctx, "stroked_glyphs");
-}
-
-#[test]
-fn skewed_glyphs() {
-    let mut ctx = get_ctx(300, 70, false);
-    let font_size: f32 = 50_f32;
-    let (font, glyphs) = layout_glyphs("Hello, world!", font_size);
-
-    ctx.set_transform(Affine::translate((0., f64::from(font_size))));
-    ctx.set_paint(REBECCA_PURPLE.with_alpha(0.5).into());
-    ctx.glyph_run(&font)
-        .font_size(font_size)
-        .glyph_transform(Affine::skew(-20_f64.to_radians().tan(), 0.0))
-        .fill_glyphs(glyphs.into_iter());
-
-    check_ref(&ctx, "skewed_glyphs");
-}
-
-#[test]
-fn scaled_glyphs() {
-    let mut ctx = get_ctx(150, 125, false);
-    let font_size: f32 = 25_f32;
-    let (font, glyphs) = layout_glyphs("Hello,\nworld!", font_size);
-
-    ctx.set_transform(Affine::translate((0., f64::from(font_size))).then_scale(2.0));
-    ctx.set_paint(REBECCA_PURPLE.with_alpha(0.5).into());
-    ctx.glyph_run(&font)
-        .font_size(font_size)
-        .fill_glyphs(glyphs.into_iter());
-
-    check_ref(&ctx, "scaled_glyphs");
-}
-
-/// ***DO NOT USE THIS OUTSIDE OF THESE TESTS***
-///
-/// This function is used for _TESTING PURPOSES ONLY_. If you need to layout and shape
-/// text for your application, use a proper text shaping library like `Parley`.
-///
-/// We use this function as a convenience for testing; to get some glyphs shaped and laid
-/// out in a small amount of code without having to go through the trouble of setting up a
-/// full text layout pipeline, which you absolutely should do in application code.
-fn layout_glyphs(text: &str, font_size: f32) -> (Font, Vec<Glyph>) {
-    const ROBOTO_FONT: &[u8] = include_bytes!("../../../examples/assets/roboto/Roboto-Regular.ttf");
-    let font = Font::new(Blob::new(Arc::new(ROBOTO_FONT)), 0);
-
-    let font_ref = {
-        let file_ref = FileRef::new(font.data.as_ref()).unwrap();
-        match file_ref {
-            FileRef::Font(f) => f,
-            FileRef::Collection(collection) => collection.get(font.index).unwrap(),
-        }
-    };
-    let font_size = skrifa::instance::Size::new(font_size);
-    let axes = font_ref.axes();
-    let variations: Vec<(&str, f32)> = vec![];
-    let var_loc = axes.location(variations.as_slice());
-    let charmap = font_ref.charmap();
-    let metrics = font_ref.metrics(font_size, &var_loc);
-    let line_height = metrics.ascent - metrics.descent + metrics.leading;
-    let glyph_metrics = font_ref.glyph_metrics(font_size, &var_loc);
-
-    let mut pen_x = 0_f32;
-    let mut pen_y = 0_f32;
-
-    let glyphs = text
-        .chars()
-        .filter_map(|ch| {
-            if ch == '\n' {
-                pen_y += line_height;
-                pen_x = 0.0;
-                return None;
-            }
-            let gid = charmap.map(ch).unwrap_or_default();
-            let advance = glyph_metrics.advance_width(gid).unwrap_or_default();
-            let x = pen_x;
-            pen_x += advance;
-            Some(Glyph {
-                id: gid.to_u32(),
-                x,
-                y: pen_y,
-            })
-        })
-        .collect::<Vec<_>>();
-
-    (font, glyphs)
-}
-
-#[test]
 fn oversized_star() {
     let mut ctx = get_ctx(100, 100, true);
 
@@ -676,5 +563,3 @@ macro_rules! compose_impl {
 fn compose_solid_src_over() {
     compose_impl!(Compose::SrcOver, "compose_solid_src_over");
 }
-
-
