+++ sparse_strips/vello_cpu/tests/clip.rs
@@ -1,9 +1,9 @@
+use crate::util::{check_ref, circular_star, crossed_line_star, get_ctx};
 use std::f64::consts::PI;
 use vello_common::color::palette::css::{DARK_BLUE, DARK_GREEN, REBECCA_PURPLE};
 use vello_common::kurbo::{Affine, BezPath, Circle, Point, Rect, Shape, Stroke};
 use vello_common::peniko::Fill;
 use vello_cpu::RenderContext;
-use crate::util::{check_ref, circular_star, crossed_line_star, get_ctx};
 
 #[test]
 fn clip_triangle_with_star() {
